Add assertions for txOpts and methodOpts
