+++ packages/contract-wrappers/src/schemas/method_opts_schema.ts
@@ -1,0 +1,7 @@
+export const methodOptsSchema = {
+    id: '/MethodOpts',
+    properties: {
+        defaultBlock: { $ref: '/BlockParam' },
+    },
+    type: 'object',
+};
