+++ packages/contract-wrappers/src/schemas/tx_opts_schema.ts
@@ -1,0 +1,8 @@
+export const txOptsSchema = {
+    id: '/TxOpts',
+    properties: {
+        gasPrice: { $ref: '/Number' },
+        gasLimit: { type: 'number' },
+    },
+    type: 'object',
+};
